libfstt: Mark the two muldiv() inline assembler outputs as early clobbers
If for whatever reason, the compiler knows that c is equal to a or b,
it may use the same register and the result would be unpredictable.
For instance, the muldiv(x, y, x) compiled with Clang can return 1
independently of y. Fortunately, we expect such case to not happen so
often. But it is easily patched by declaring the two outputs as early
clobbers.

Co-Author: Sébastien Bardin <sebastien.bardin@cea.fr>
Co-Author: Richard Bonichon <richard.bonichon@gmail.com>
Signed-off-by: Guillem Jover <guillem@hadrons.org>